+++ WatchFriends/app/src/main/java/nmct/jaspernielsmichielhein/watchfriends/helper/Utils.java
@@ -1,6 +1,9 @@
 package nmct.jaspernielsmichielhein.watchfriends.helper;
 
+import android.content.Context;
+import android.content.DialogInterface;
 import android.graphics.Typeface;
+import android.support.v7.app.AlertDialog;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ListAdapter;
@@ -8,6 +11,9 @@ import android.widget.ListView;
 import android.widget.TextView;
 
 import com.google.android.gms.common.SignInButton;
+
+import nmct.jaspernielsmichielhein.watchfriends.R;
+import permissions.dispatcher.PermissionRequest;
 
 public class Utils {
 
@@ -43,6 +49,26 @@ public class Utils {
                 return;
             }
         }
+    }
+
+    public static void showRationaleDialog(Context context, String message, final PermissionRequest request) {
+        new AlertDialog.Builder(context, R.style.customDialog)
+                .setPositiveButton("Allow", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialogInterface, int i) {
+                        request.proceed();
+                    }
+                })
+                .setNegativeButton("Deny", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialogInterface, int i) {
+                        request.cancel();
+                    }
+                })
+                .setCancelable(false)
+                .setTitle("Permission denied")
+                .setMessage(message)
+                .show();
     }
 
 }
